requirements.txt
streamlit==1.38.0
pillow==10.4.0
numpy==1.26.4